requirements.txt
pycryptodome==3.20.0
matplotlib==3.8.2